11월 04일 작업
일반 공격
=> 바람 기본 공격: 단일 피해 공격력해당 퍼센트(완료)

=> 불 기본 공격 :  범위 피해  공격력 해당 퍼센트(완료)

=> 물 기본 공격 : 피해는 일반 공격력이지만 이동속도를 낮춤(완료)

서브스킬 작업
=> 윈드 드릴(바람 서브 스킬): 단일 피해 공격력 * 해당 퍼센트 총 6회의 피해를 줌(완료)

=> 파이어볼(불 서브 스킬): 관통 피해 공격력* 해당 퍼센트(완료)

=> 워터베리어(물 서브 스킬): 보호막 막아지는 횟수 1번(완료)

-------------------------------------------

[버그- 계속 수정하고 테스트 해봐야함 위의 문제 확인해보고 추가 수정할 거 기입]

사운드 증폭시킬것(완료)

플레이어 이동 화살표로 변경하고 공격을 asd로 변경(완료)

스킬 선택할때 스킬 선택 번호를 기입(완료)

일시정지 버튼 esc 추가(완료)

캐릭터 정보 단축키: tab(완료)

플레이어 일반 공격 : 자동으로 나가게 설정(완료)

보스 피격 이펙트 조정(완료)

맵 기믹인 토네이도 커브드 월드랑 합쳐서 바꾸기 <= 작업 끝나는데로 작업들어가기(완료)

스테이지 브금 루프가 어색함. 조정 필요(완료)

a,d 혹은 ←, → 눌렀을 때 넘어가게 하는 걸 추가(완료)

레벨업 스킬 선택창에 1, 2, 3번 기입이 안 되어있음(완료)

보스 등장 시 브금 교체(완료)

보스 피격 시 깜빡이는 거 어색함. <= 간격 수정(완료)

폰트 깨짐 수정(완료)

추가: 스킬작업하느라 몬스터에 맞아도 피격판정이 안뜸(완료)

보스 피격 함수도 한번 체크할 것(완료)

서브 스킬 쿨타임 카운트 UI 연결(완료)

기획자님이 올려주신 버그, 피드백 리스트 링크도 참고할 것

+++ Assets/10.DataTable/DB_StatusEntity.cs
@@ -104,7 +104,7 @@ public class DB_StatusEntity_04
 
     public int weaponType; // 서브 스킬 무기 속성
 
-    public float typeLevel; // 서브 스킬의 무기 레벨
+    public int typeLevel; // 서브 스킬의 무기 레벨
 
     public string name; // 서브 스킬 이름
 
@@ -118,7 +118,7 @@ public class DB_StatusEntity_04
 
     public float penetDamage; // 불 서브 스킬의 관통 데미지
 
-    public float penetCount; // 불 서브 스킬의 최대 맞는 개체수(얼만큼 관통시킬지.... 기획상으론 끝까지이지만 일단 데이터 기입)
+    public int penetCount; // 불 서브 스킬의 최대 맞는 개체수(얼만큼 관통시킬지.... 기획상으론 끝까지이지만 일단 데이터 기입)
 
     public float barrierDuration; // 물 서브 스킬 베리어의 지속시간(시간을 표현해야함으로 float으로 선언)
